requirements.txt
fastapi==0.88.0
iteration_utilities==0.11.0
mariadb==1.1.5.post2
numba==0.56.4
numpy==1.23.5
pygame==2.1.2
python_dateutil==2.8.2
requests==2.28.1
scipy==1.9.3
tqdm==4.64.1
xmltodict==0.13.0
